microservice/MemeGeneratorExecutor3909942/0_requests_pillow/v1/requirements.txt
jina==3.15.1.dev14
docarray==0.21.0
openai==0.27.5
pytest
requests==2.26.0
Pillow==8.3.2